// src/API.ts
export type Todo = {
  __typename?: 'Todo'
  id?: string
  name: string
  description?: string | null
  createdAt?: string
  updatedAt?: string
} | null
// ...
export type ModelTodoConnection = {
  __typename: 'ModelTodoConnection'
  items?: Array<Todo | null> | null
  nextToken?: string | null
}
// ...
export type ListTodosQuery = {
  listTodos?: {
    __typename: 'ModelTodoConnection'
    items?: Array<{
      __typename: 'Todo'
      id: string
      name: string
      description?: string | null
      createdAt: string
      updatedAt: string
    } | null> | null
    nextToken?: string | null
  } | null
}